# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CodiceCivile.ai API
  description: |
    REST API for the CodiceCivile.ai platform providing semantic search,
    AI-powered chat, and legal article management.

    **Features:**
    - Semantic search across Italian Civil Code articles
    - AI-powered legal assistance with Claude
    - User bookmarks and search history
    - Vector embeddings for similarity search
  version: 1.0.0
  contact:
    name: CodiceCivile.ai Support
    email: support@codicecivile.ai
  license:
    name: Proprietary
servers:
  - url: https://api.codicecivile.ai/v1
    description: Production server
  - url: https://staging-api.codicecivile.ai/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Local development server
...
paths:
...
  /search:
    post:
      tags: [search]
      summary: Semantic search
      description: |
        Perform semantic search across legal articles using natural language queries.
        Uses vector embeddings for similarity-based ranking.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  description: Natural language search query
                  example: "diritti del consumatore nei contratti"
                filters:
                  type: object
                  properties:
                    book:
                      type: integer
                      minimum: 1
                      maximum: 6
                    article_type:
                      type: string
                      enum: [contract, property, family, succession, obligations]
                    date_range:
                      type: object
                      properties:
                        start:
                          type: string
                          format: date
                        end:
                          type: string
                          format: date
                limit:
                  type: integer
                  default: 10
                  minimum: 1
                  maximum: 50
                sort_by:
                  type: string
                  enum: [relevance, date, article_number]
                  default: relevance
      responses:
        '200':
          description: Successful search
          content:
            application/json:
              schema:
                type: object
                properties:
                  query:
                    type: string
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        article:
                          $ref: '#/components/schemas/Article'
                        relevance_score:
                          type: number
                          format: float
                          minimum: 0
                          maximum: 1
                        matched_snippet:
                          type: string
                          description: Relevant excerpt from article content
                  total_results:
                    type: integer
                  search_time_ms:
                    type: integer
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # CHAT
  /chat:
    post:
      tags: [chat]
      summary: Chat with AI assistant
      description: |
        Send messages to the AI legal assistant powered by Claude.
        Supports streaming responses and article context.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - messages
              properties:
                messages:
                  type: array
                  items:
                    type: object
                    required:
                      - role
                      - content
                    properties:
                      role:
                        type: string
                        enum: [user, assistant]
                      content:
                        type: string
                context:
                  type: object
                  description: Optional article context for the chat
                  properties:
                    article_id:
                      type: string
                    article_title:
                      type: string
                    article_content:
                      type: string
                stream:
                  type: boolean
                  default: false
                  description: Enable streaming responses
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  role:
                    type: string
                    enum: [assistant]
                  content:
                    type: string
                  related_articles:
                    type: array
                    items:
                      type: string
                      description: Article IDs cited in the response
                  timestamp:
                    type: string
                    format: date-time
            text/event-stream:
              schema:
                type: string
                description: Server-sent events for streaming responses
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # EMBEDDINGS
...
components:
  schemas:
    Article:
      type: object
      properties:
        id:
          type: string
          example: "art-1234"
        article_number:
          type: string
          example: "1234"
        book:
          type: integer
          minimum: 1
          maximum: 6
          example: 4
        title:
          type: string
          example: "Diritti del consumatore"
        content:
          type: string
          example: "Il consumatore ha diritto..."
        keywords:
          type: array
          items:
            type: string
          example: ["consumatore", "contratti", "diritti"]
        last_modified:
          type: string
          format: date
          example: "2024-01-15"
        embedding:
          type: array
          items:
            type: number
            format: float
          description: 768-dimensional vector embedding (optional)
          nullable: true
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: "INVALID_REQUEST"
            message:
              type: string
              example: "Invalid article ID provided"
            details:
              type: object
              nullable: true
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication endpoint
security:
  - BearerAuth: []